finished

+++ server/index.js
@@ -9,7 +9,7 @@ const {
   fetchFavorites,
   destroyFavorite,
   authenticate,
-  findUserWithToken
+  findUserByToken
 } = require('./db');
 const express = require('express');
 const app = express();
@@ -20,6 +20,15 @@ const path = require('path');
 app.get('/', (req, res)=> res.sendFile(path.join(__dirname, '../client/dist/index.html')));
 app.use('/assets', express.static(path.join(__dirname, '../client/dist/assets'))); 
 
+const isLoggedIn = async(req, res, next)=> {
+  try {
+    req.user = await findUserByToken(req.headers.authorization);
+    next();
+  }
+  catch(ex){
+    next(ex);
+  }
+};
 
 app.post('/api/auth/login', async(req, res, next)=> {
   try {
@@ -30,9 +39,9 @@ app.post('/api/auth/login', async(req, res, next)=> {
   }
 });
 
-app.get('/api/auth/me', async(req, res, next)=> {
+app.get('/api/auth/me', isLoggedIn, (req, res, next)=> {
   try {
-    res.send(await findUserWithToken(req.headers.authorization));
+    res.send(req.users);
   }
   catch(ex){
     next(ex);
@@ -48,26 +57,41 @@ app.get('/api/users', async(req, res, next)=> {
   }
 });
 
-app.get('/api/users/:id/favorites', async(req, res, next)=> {
+app.get('/api/users/:id/favorites', isLoggedIn, async(req, res, next)=> {
   try {
-    res.send(await fetchFavorites(req.params.id));
+    if(req.params.userId !== req.user.id){
+      const error = Error('not authorized');
+      error.status = 401;
+      throw error;
+    }
+    res.send(await fetchUserSkills(req.params.id));
   }
   catch(ex){
     next(ex);
   }
 });
 
-app.post('/api/users/:id/favorites', async(req, res, next)=> {
+app.post('/api/users/:id/favorites', isLoggedIn, async(req, res, next)=> {
   try {
-    res.status(201).send(await createFavorite({ user_id: req.params.id, product_id: req.body.product_id}));
+    if(req.params.id !== req.user.id){
+      const error = Error('not authorized');
+      error.status = 401;
+      throw error;
+    }
+  res.status(201).send(await createFavorite({user_id: req.params.id, skill_id: req.body.skill_id}));
   }
   catch(ex){
     next(ex);
   }
 });
 
-app.delete('/api/users/:user_id/favorites/:id', async(req, res, next)=> {
+app.delete('/api/users/:user_id/favorites/:id', isLoggedIn, async(req, res, next)=> {
   try {
+    if(req.params.userId !== req.user.id){
+      const error = Error('not authorized');
+      error.status = 401;
+      throw error;
+    }
     await destroyFavorite({user_id: req.params.user_id, id: req.params.id });
     res.sendStatus(204);
   }
